refactor: convert `GameStats` to Typescript
The main `GameStats` class has been rewritten in Typescript and reorganized.
This eliminates the need for a separately maintained `.d.ts` file, while also
making it easier for others to contribute in the future.

I've also removed Rollup in favor of `tsup`. This simplifies the build
pipeline significantly.

BREAKING CHANGE: Exported files are moved and renamed.